Migration des modeles vers l'app Manager
